sys-libs/safeclib: do not remove manpages if not built
Closes: https://bugs.gentoo.org/784686
Package-Manager: Portage-3.0.18, Repoman-3.0.3

+++ sys-libs/safeclib/safeclib-3.6.0_p20210407-r2.ebuild
@@ -87,12 +87,15 @@ src_install() {
 	cd "${S}-lib" || die
 	default
 	einstalldocs
-	use doc && dodoc -r doc/.
 
-	# wcsstr towupper towlower manpages collide with sys-apps/man-pages
-	rm "${ED}/usr/share/man/man3/towlower.3" || die
-	rm "${ED}/usr/share/man/man3/towupper.3" || die
-	rm "${ED}/usr/share/man/man3/wcsstr.3" || die
+	if use doc ; then
+		dodoc -r doc/.
+
+		# wcsstr towupper towlower manpages collide with sys-apps/man-pages
+		rm "${ED}/usr/share/man/man3/towlower.3" || die
+		rm "${ED}/usr/share/man/man3/towupper.3" || die
+		rm "${ED}/usr/share/man/man3/wcsstr.3" || die
+	fi
 }
 
 src_test() {
